Fix lotto_improved: RNDINT type conversion
Fixed RNDINT calls to use INT() conversion for Float→Int.
Lotto program now ready to test (SUB modifies arrays!)

Status: 38/46 tests passing (83%)

+++ examples/lotto_improved.bas
@@ -30,7 +30,7 @@ ENDFUNCTION
 SUB generateUniqueNumbers(arr, count, minVal, maxVal)
     LET i = 0.0
     WHILE i < count
-        LET num = RNDINT(minVal, maxVal)
+        LET num = INT(RNDINT(INT(minVal), INT(maxVal)))
         
         IF hasDuplicate(arr, i, num) == 0.0 THEN
             LET arr(i) = num
